backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openai==1.10.0
groq>=0.33.0
reportlab==4.0.8
python-docx==1.1.0
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.1.0
markdown==3.5.1
html2text==2024.2.26
PyPDF2==3.0.1
pdfplumber==0.10.3
xhtml2pdf==0.2.15

